Move Error class to common package and add InternalServerError class
