+++ src/main/scala/common/errors/Error.scala
@@ -1,0 +1,3 @@
+package common.errors
+
+class Error extends Exception
